+++ lib/sidekiq_adhoc_job/services/schedule_adhoc_job.rb
@@ -9,6 +9,9 @@ module SidekiqAdhocJob
         acc
       end
       @worker_klass = WorkerClassesLoader.find_worker_klass(job_name)
+      @worker_klass_inspector = Utils::ClassInspector.new(worker_klass)
+      @allowed_params = worker_klass_inspector.required_parameters(:perform) + worker_klass_inspector.optional_parameters(:perform)
+
       parse_params
     end
 
@@ -18,10 +21,10 @@ module SidekiqAdhocJob
 
     private
 
-    attr_reader :request_params, :worker_klass, :allowed_params, :worker_params
+    attr_reader :request_params, :worker_klass, :worker_klass_inspector,
+                :allowed_params, :worker_params
 
     def parse_params
-      @allowed_params = worker_klass.new.method(:perform).parameters.reject { |type, _| type == :rest }.flat_map(&:last)
       @worker_params = allowed_params.map { |key| StringUtil.parse(request_params[key]) }
       if !!request_params[:rest_args] && !request_params[:rest_args].empty?
         @worker_params += request_params[:rest_args].split(',').map { |arg| StringUtil.parse(arg.strip) }